==> map_spatial_cv_abun/step4_test_model.R <==
#step4:test model###############################################################
rm(list = ls())
library(caret)
library(randomForest)
library(ggplot2)
require(ggplot2)
require(caret)
require(randomForest)
require(future.apply)
require(parallel)
require(doParallel)
require(foreach)
library(blockCV)

load(file = 'pathogen/map_spatial_cv_abun/optvar.Rdata')
load(file = 'pathogen/map_spatial_cv_abun/ml_df.Rdata')
dat <- df_filter
dat2 <- na.omit(dat)
dat3 <- dat2[,-1]
x <- dat3[,optvar]   
logBBB <- dat2$mean_otu_abun

i <- 1

set.seed(i)
load(file = 'pathogen/map_spatial_cv_abun/inTrain.Rdata')
trainx <- x[inTrain, ]
testx <- x[-inTrain, ]
trainy <- logBBB[inTrain]
testy <- logBBB[-inTrain]

require(ggplot2)
library(data.table)

final_grid <- data.table()

load("pathogen/map_spatial_cv_abun/HPB-500_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(500,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-800_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(800,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-1000_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(1000,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-1200_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(1200,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-1500_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(1500,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-2000_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(2000,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

load("pathogen/map_spatial_cv_abun/HPB-2500_abun.Rdata")
grid_search <- data.table(custom$results)
grid_search$'ntree' <- rep(2500,nrow(grid_search))
final_grid <- dplyr::bind_rows(final_grid,grid_search)

write.csv(final_grid,file = 'pathogen/map_spatial_cv_abun/grid_cv_abun.csv',row.names = F)
final_grid$ntree[which.max(final_grid$Rsquared)]
final_grid$ntree[which.min(final_grid$RMSE)]

load("pathogen/map_spatial_cv_abun/HPB-1500_abun.Rdata")
custom$finalModel

plot <- data.frame(custom$pred$pred,custom$pred$obs)
write.csv(plot,file = 'pathogen/map_spatial_cv_abun/train_cv_abun.csv',row.names = F)

predictRF <- predict(custom$finalModel,newdata = testx)

rsq <- function(x, y) summary(lm(y~x))$r.squared
rsq(testy, predictRF)
R2(testy, predictRF)

plot <- data.frame(predictRF,testy)
write.csv(plot,file = 'pathogen/map_spatial_cv_abun/test_cv_abun.csv',row.names = F)
